% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_dispersion}
\alias{extract_dispersion}
\title{Extract dispersion/precision parameter (on natural scale) heuristically for many glmmTMB families}
\usage{
extract_dispersion(model)
}
\description{
\itemize{
\item Negative binomial (nbinom2): phi or theta (size) often in fixef(model)$disp on log scale -> exp()
\item Beta: precision phi often in fixef(model)$disp on log scale -> exp()
}
}
