% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{run_one_sim}
\alias{run_one_sim}
\title{Single-simulation function}
\usage{
run_one_sim(sim_i, lambda, n_obs)
}
\value{
A named list with estimates for poisson, nbinom and beta-backtransformed
}
\description{
Single-simulation function
}
