% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{safe_glmmTMB_fit}
\alias{safe_glmmTMB_fit}
\title{safe glmmTMB fitting and parameter extraction}
\usage{
safe_glmmTMB_fit(formula, data, family)
}
\description{
safe glmmTMB fitting and parameter extraction
}
