% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruvdists-package.R
\docType{package}
\name{bruvdists-package}
\alias{bruvdists}
\alias{bruvdists-package}
\title{bruvdists: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beckyfisher/bruvdists}
  \item Report bugs at \url{https://github.com/beckyfisher/bruvdists/issues}
}

}
\author{
\strong{Maintainer}: Rebecca Fisher \email{r.fisher@aims.gov.au} (\href{https://orcid.org/0000-0001-5148-6731}{ORCID}) [copyright holder]

}
\keyword{internal}
